requirements.txt
cupy-cuda100==7.1.0
cycler==0.10.0
fastrlock==0.4
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.13.0
